Inverse the hessian and multiply it by the score

Inverse the hessian and multiply it by the score

hessian_solver(par, XX_band, Xy, pen, w, diff)

Arguments

par
The parameter vector

XX_band
The matrix \(X^T X\) where X is the design matrix. This argument is given in the form of a band matrix, i.e., successive columns represent superdiagonals.

Xy
The vector of currently estimated points \(X^T y\), where \(y\) is the y-coordinate of the data.

pen
Positive penalty constant.

w
Vector of weights. Has to be of length

diff
The order of the differences of the parameter. Equals degree + 1 in adaptive spline regression.

Value

The solution of the linear system: $$(X^T X + pen D^T diag(w) D) ^ {-1} X^T y - par$$